Created new (empty) user profile screen
