+++ app/components/NavBar/Navbar.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Card, Text, Header } from 'react-native-elements';
 import { Button, StyleSheet, View, TouchableHighlight, Image, ScrollView } from 'react-native';
-
+import styles from './styles'
 
 export default class Navbar extends Component<{}> {
 
@@ -9,45 +9,50 @@ export default class Navbar extends Component<{}> {
         super(props);
     }
 
-    _onPressButton() {
+    onPressButton() {
         console.log('hi')
     }
 
     render() {
         return (
             <View style={styles.container}>
+
                 <View style={styles.button} >
                     <TouchableHighlight onPress={() => this.props.navigation.navigate('HomePage')}>
                         <Image
-                            source={require('../../images/home.png')}
+                            source={require('../../../images/home.png')}
                         />
                     </TouchableHighlight>
                 </View>
+
                 <View style={styles.button} >
                     <TouchableHighlight onPress={() => this.props.navigation.navigate('ExperimentBrowse')}>
                         <Image
-                            source={require('../../images/experiments.png')}
+                            source={require('../../../images/experiments.png')}
                         />
                     </TouchableHighlight>
                 </View>
+
                 <View style={styles.biggerButton} >
                     <TouchableHighlight onPress={this._onPressButton}>
                         <Image
-                            source={require('../../images/completeSurvey.png')}
+                            source={require('../../../images/completeSurvey.png')}
                         />
                     </TouchableHighlight>
                 </View>
+
                 <View style={styles.button} >
                     <TouchableHighlight onPress={this._onPressButton}>
                         <Image
-                            source={require('../../images/stats.png')}
+                            source={require('../../../images/stats.png')}
                         />
                     </TouchableHighlight>
                 </View>
+
                 <View style={styles.button} >
-                    <TouchableHighlight onPress={this._onPressButton}>
+                    <TouchableHighlight onPress={() => this.props.navigation.navigate('UserProfile')}>
                         <Image
-                            source={require('../../images/userProfile.png')}
+                            source={require('../../../images/userProfile.png')}
                         />
                     </TouchableHighlight>
                 </View>
@@ -55,25 +60,5 @@ export default class Navbar extends Component<{}> {
         );
     }
 }
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        flexDirection: 'row',
-        justifyContent: 'center',
-        alignItems: 'baseline',
-        marginBottom: 80,
-        marginTop: 10,
-        marginLeft: 15,
-    },
-    button: {
-        width: 60,
-        height: 60,
-    },
-    biggerButton: {
-        width:90,
-        height: 100,
-    },
-});
 
 module.exports = Navbar;
