+++ app/components/NavBar/styles.js
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+    container: {
+        flex: 1,
+        flexDirection: 'row',
+        justifyContent: 'center',
+        alignItems: 'baseline',
+        marginBottom: 80,
+        marginTop: 10,
+        marginLeft: 15,
+    },
+    button: {
+        width: 60,
+        height: 60,
+    },
+    biggerButton: {
+        width:90,
+        height: 100,
+    },
+});
