+++ app/components/ReactionModal/styles.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+    dividerTextStyle: {
+        textAlign: 'center',
+        fontSize: 20,
+    },
+    emojiDivider: {
+        flex: 1,
+        flexDirection: 'row',
+        justifyContent: 'center',
+        marginTop: 100
+    },
+    emojiToken: {
+        width: 100,
+        height: 100,
+        margin: 7
+},});
