+++ app/config/navigationNames.js
@@ -1,0 +1,7 @@
+import UserProfile from "../screens/UserProfile";
+
+export default class Screens {
+    static UserProfile = "UserProfile";
+    static Home = "HomePage";
+    static Experiments = "Experiment"
+}
