+++ app/config/styles.js
@@ -1,0 +1,1 @@
+//global styles should go here
