+++ app/screens/HomePage.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import {Icon,Button, Text, Tile, SearchBar, List, ListItem, Card} from 'react-native-elements';
 import {ScrollView, Modal, StyleSheet} from 'react-native';
 import {View} from 'react-native';
-import ReactionModal from '../components/ReactionModal.js'
-import Navbar from '../components/Navbar.js'
+import ReactionModal from '../components/ReactionModal/ReactionModal.js'
+import Navbar from '../components/NavBar/Navbar.js'
 
 
 export default class HomePage extends Component<{}> {
@@ -21,15 +21,12 @@ export default class HomePage extends Component<{}> {
     handleChange(event) {
         this.setState({ searchText: event});
     }
+
     handleClickExperiment(event){
         console.log("Clicked List item "+event.toString())
     }
 
-
     render() {
-
-
-
         //Temporary list...
         const list = [
             {
@@ -52,8 +49,7 @@ export default class HomePage extends Component<{}> {
                 subtitle: 'Veggies and Keenwah man',
                 icon: 'av-timer'
             }
-        ]
-
+        ];
 
         return (
             <View style={{flex: 1}}>
@@ -63,12 +59,10 @@ export default class HomePage extends Component<{}> {
                     <Text h4> Experiments </Text>
                     <Text> See how you are doing with your experiments </Text>
 
-
                     <Card title="Day 2"
                           titleStyle = {styles.dividerTextStyle}>
 
                     </Card>
-
                     <Card >
                         <View style={{flex: 1, flexDirection: 'row', justifyContent: 'space-between'}}>
                             <Text style = {styles.dividerTextStyle}> Daily Checkup </Text>
@@ -87,16 +81,11 @@ export default class HomePage extends Component<{}> {
                                 ))
                             }
                         </List>
-
                     </Card>
-
                     <Card title="Goal Streak"
                           titleStyle = {styles.dividerTextStyle}>
 
                     </Card>
-
-
-
                 </View>
 
             </ScrollView>
@@ -107,7 +96,7 @@ export default class HomePage extends Component<{}> {
         )}
 }
 
-module.exports = HomePage
+module.exports = HomePage;
 
 const styles = StyleSheet.create({
     dividerTextStyle: {
